# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartSpend API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Dashboard summary for a month
      parameters:
        - name: month
          in: query
          schema:
            type: string
          description: "YYYY-MM, defaults to current month"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /analytics/monthly:
    get:
      operationId: getMonthlyAnalytics
      tags: [analytics]
      summary: Monthly income vs expense trends
      parameters:
        - name: months
          in: query
          schema:
            type: integer
          description: "Number of past months (default 6)"
      responses:
        "200":
          description: Monthly trends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"
  /analytics/categories:
    get:
      operationId: getCategoryAnalytics
      tags: [analytics]
      summary: Spending by category for a month
      parameters:
        - name: month
          in: query
          schema:
            type: string
          description: "YYYY-MM, defaults to current month"
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryStat"
  /ai/query:
    post:
      operationId: aiQuery
      tags: [ai]
      summary: Natural language query about spending
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIQueryInput"
      responses:
        "200":
          description: AI answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIQueryResponse"
  /ai/categorize:
    post:
      operationId: aiCategorize
      tags: [ai]
      summary: Auto-categorize a transaction description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AICategorizationInput"
      responses:
        "200":
          description: Suggested category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AICategorizationResponse"
  /ai/summary:
    get:
      operationId: getAISummary
      tags: [ai]
      summary: AI-generated monthly financial summary
      parameters:
        - name: month
          in: query
          schema:
            type: string
          description: "YYYY-MM, defaults to current month"
      responses:
        "200":
          description: AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AISummaryResponse"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - createdAt
    UserInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 1
      required:
        - email
        - password
        - name
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
...
    DashboardSummary:
      type: object
      properties:
        totalIncome:
          type: number
        totalExpense:
          type: number
        savings:
          type: number
        savingsRate:
          type: number
        month:
          type: string
        transactionCount:
          type: integer
        topCategories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryStat"
      required:
        - totalIncome
        - totalExpense
        - savings
        - savingsRate
        - month
        - transactionCount
        - topCategories
    MonthlyTrend:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        expense:
          type: number
        savings:
          type: number
      required:
        - month
        - income
        - expense
        - savings
    CategoryStat:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
        count:
          type: integer
        icon:
          type: string
        color:
          type: string
      required:
        - category
        - amount
        - percentage
        - count
        - icon
        - color
    AIQueryInput:
      type: object
      properties:
        query:
          type: string
          minLength: 1
        month:
          type: string
          description: "Optional YYYY-MM context"
      required:
        - query
    AIQueryResponse:
      type: object
      properties:
        answer:
          type: string
        dataUsed:
          type: string
      required:
        - answer
        - dataUsed
    AICategorizationInput:
      type: object
      properties:
        description:
          type: string
          minLength: 1
        amount:
          type: number
      required:
        - description
    AICategorizationResponse:
      type: object
      properties:
        category:
          type: string
        confidence:
          type: number
        reasoning:
          type: string
      required:
        - category
        - confidence
        - reasoning
    AISummaryResponse:
      type: object
      properties:
        summary:
          type: string
        insights:
          type: array
          items:
            type: string
        month:
          type: string
        generatedAt:
          type: string
          format: date-time
      required:
        - summary
        - insights
        - month
        - generatedAt